requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pytest==7.4.2
Faker==18.6.2
factory_boy
behave